// shared/models/user.model.ts
export class User {
  _id?: string;
  username?: string;
  password?:string;
  email?: string;
  title: string;
  role?: string;
  gender?: string;
  firstName: string;
  lastName: string;
  address: string;
  state: string;
  city: string;
  age: number;
  birthDate: any;
  purposeOfVisit?: string;
  mobileNo: string;
  associated_agency?:string;
  deleted:boolean;
  selected?:boolean;
  markup:number=0;

}